requirements.txt
python-dotenv
streamlit
requests 